detail.js에서 movieInfo로 js분리 및 import 적용
