+++ js/mainBanner.js
@@ -68,7 +68,7 @@ const optionsBanner = {
       const url = new URL("http://127.0.0.1:5502/detail.html"); //로컬주소라 추후 도메인변경해야함
       detailBtn[index].addEventListener('click', (e) => {
         const detailUrl = `${url.href}?id=${idx.id}`
-        window.open(detailUrl, '_blank'); 
+        window.open(detailUrl, '_self'); 
 
         console.log(`클릭클릭 : ${index}${idx.id}${detailUrl}`)
       })
