+++ js/upcomMov.js
@@ -36,7 +36,7 @@ export function loadUpcomingCards(location, options, arr1, arr2) {
         <p class="rating">rating : ${object.vote_average}</p>
       </div>
       <div class="detailBtnWrap">
-        <button onclick="location.href='detail.html?id=${object.id}'" class="goDetail">상세보기</button>
+        <button onclick="location.href='./detail.html?id=${object.id}'" class="goDetail">상세보기</button>
       </div>
       `;
